refactor: Verctor 패키지 이동
