+++ src/main/java/janggi/domain/Vector.java
@@ -1,0 +1,24 @@
+package janggi.domain;
+
+import janggi.domain.piece.Side;
+import java.util.List;
+
+public record Vector(int y, int x) {
+
+    public Vector side(Side side) {
+        if (side == Side.CHO) {
+            return new Vector(-y, x);
+
+        }
+        return new Vector(y, x);
+    }
+
+    public static List<List<Vector>> rotate(List<List<Vector>> vectorsList) {
+        return vectorsList.stream()
+                .map(list -> list.stream()
+                        .map(vector -> new Vector(vector.x(), -vector.y())) // (x, y) → (y, -x) 변환
+                        .toList()
+                )
+                .toList();
+    }
+}
